requirements.txt
python-dotenv
sw_mc_lib